Consolidate recievers under single dir
