+++ src/receivers/renderer.ts
@@ -1,10 +1,10 @@
 import * as VF from 'vexflow';
 import { BoundingBox, Clef, StaveModifierPosition } from 'vexflow';
-import { Attributes } from './attributes';
-import { CodePrinter } from './codeprinter';
+import { Attributes } from '../attributes';
+import { CodePrinter } from '../codeprinter';
+import { Producer } from '../producers/producer';
+import { CodeTracker, NoteMessage, VexmlMessage, VexmlMessageReceiver } from '../types';
 import { Notations } from './notations';
-import { Producer } from './producers/producer';
-import { CodeTracker, NoteMessage, VexmlMessage, VexmlMessageReceiver } from './types';
 
 export type RendererOptions = {
   codeTracker?: CodeTracker;
